format that sh*t

+++ contracts/mock/RevenueToken4626.sol
@@ -3,9 +3,8 @@ pragma solidity 0.8.9;
 import "./RevenueToken.sol";
 
 contract RevenueToken4626 is RevenueToken {
-
     address private _asset;
-    uint private _multiplier;
+    uint256 private _multiplier;
 
     constructor(address assetAddr) {
         _asset = assetAddr;
@@ -16,17 +15,17 @@ contract RevenueToken4626 is RevenueToken {
         _asset = assetAddr;
     }
 
-    function setAssetMultiplier(uint multiplier) public {
+    function setAssetMultiplier(uint256 multiplier) public {
         _multiplier = multiplier;
     }
 
     // mimic eip-4626
-    function asset() public view returns(address) {
+    function asset() public view returns (address) {
         return _asset;
     }
 
     // mimic eip-4626
-    function previewRedeem(uint256 amount) public view returns(uint) {
+    function previewRedeem(uint256 amount) public view returns (uint256) {
         return amount * _multiplier;
     }
 }
